+++ src/main/java/com/epam/reportportal/annotations/attribute/MultiKeyAttribute.java
@@ -16,19 +16,18 @@
 
 package com.epam.reportportal.annotations.attribute;
 
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+import java.lang.annotation.*;
 
 /**
- * Annotation used in {@link Attributes} as field, to provide multiple {@link com.epam.ta.reportportal.ws.model.attribute.ItemAttributesRQ}
- * with both 'key' and 'value' fields specified.
- * Used to prevent duplication of {@link Attribute} annotation with the same value and different keys
- *
- * @author <a href="mailto:ivan_budayeu@epam.com">Ivan Budayeu</a>
+ * Annotation used per se or in {@link Attributes} as field, to provide multiple
+ * {@link com.epam.ta.reportportal.ws.model.attribute.ItemAttributesRQ} with both 'key' and 'value' fields specified.
+ * Used to prevent duplication of {@link Attribute} annotation with the same value and different keys.
  */
+@Inherited
+@Documented
 @Retention(RetentionPolicy.RUNTIME)
-@Target({})
+@Repeatable(MultiKeyAttributeGroup.class)
+@Target({ ElementType.METHOD, ElementType.TYPE, ElementType.CONSTRUCTOR })
 public @interface MultiKeyAttribute {
 
 	String[] keys();