+++ src/main/java/com/epam/reportportal/annotations/attribute/MultiValueAttribute.java
@@ -16,19 +16,18 @@
 
 package com.epam.reportportal.annotations.attribute;
 
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+import java.lang.annotation.*;
 
 /**
- * Annotation used in {@link Attributes} as field, to provide multiple {@link com.epam.ta.reportportal.ws.model.attribute.ItemAttributesRQ}
- * with both 'key' (optional) and 'value' fields specified.
- * Used to prevent duplication of {@link Attribute} annotation with the same key and different values
- *
- * @author <a href="mailto:ivan_budayeu@epam.com">Ivan Budayeu</a>
+ * Annotation used per se or in {@link Attributes} as field, to provide multiple
+ * {@link com.epam.ta.reportportal.ws.model.attribute.ItemAttributesRQ} with both 'key' (optional) and 'value' fields specified.
+ * Used to prevent duplication of {@link Attribute} annotation with the same key and different values.
  */
+@Inherited
+@Documented
 @Retention(RetentionPolicy.RUNTIME)
-@Target({})
+@Repeatable(MultiValueAttributeGroup.class)
+@Target({ ElementType.METHOD, ElementType.TYPE, ElementType.CONSTRUCTOR })
 public @interface MultiValueAttribute {
 
 	String key() default "";
@@ -36,7 +35,7 @@ public @interface MultiValueAttribute {
 	String[] values();
 
 	/**
-	 * @return 'true' if key of the resulted entity should be NULL, otherwise {@link MultiValueAttribute#key()} will be used
+	 * @return if key of the resulted entity should be {@code null}, otherwise {@link MultiValueAttribute#key()} will be used.
 	 */
 	boolean isNullKey() default false;
 }